+++ src/tr/trColor.hpp
@@ -11,10 +11,10 @@ namespace tr
 		Vector4 ToVector() const
 		{
 			return Vector4(
-				double(b),
-				double(g),
-				double(r),
-				double(a)
+				float(b),
+				float(g),
+				float(r),
+				float(a)
 			);
 		}
